refactor: remove dup stub entry object
