+++ packages/testing/src/execution_testing/benchmark/__init__.py
@@ -8,12 +8,9 @@ from .benchmark_code_generator import (
     ExtCallGenerator,
     JumpLoopGenerator,
 )
-from .stub_config import StubConfig, StubEntry
 
 __all__ = (
     "BenchmarkCodeGenerator",
     "ExtCallGenerator",
     "JumpLoopGenerator",
-    "StubConfig",
-    "StubEntry",
 )
