--- packages/testing/src/execution_testing/benchmark/tests/__init__.py
@@ -1,1 +1,0 @@
-"""Tests for the benchmark module."""
