+++ packages/testing/src/execution_testing/cli/pytest_commands/plugins/shared/tests/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for shared pytest plugin utilities."""
